updated headers

+++ src/classes/cat.h
@@ -48,7 +48,7 @@ class CatalogReader
     int read_local_planets(CelestialObject** cels, int max);
 
     protected:
-    void read_field_onebased(char* buffer, int start, int end, char* out);
+    void read_field_onebased(char* buffer, size_t start, int end, char* out);
 };
 
 #endif
